move api/common to api/browser
