+++ src/vs/workbench/api/node/extHostDiagnostics.ts
