+++ src/vs/workbench/api/node/extHostLanguageFeatures.ts
@@ -11,14 +11,14 @@ import {IDisposable, disposeAll} from 'vs/base/common/lifecycle';
 import {Remotable, IThreadService} from 'vs/platform/thread/common/thread';
 import {Range as EditorRange} from 'vs/editor/common/core/range';
 import * as vscode from 'vscode';
-import * as TypeConverters from 'vs/workbench/api/common/extHostTypeConverters';
-import {Range, DocumentHighlightKind, Disposable, Diagnostic, SignatureHelp} from 'vs/workbench/api/common/extHostTypes';
+import * as TypeConverters from 'vs/workbench/api/node/extHostTypeConverters';
+import {Range, DocumentHighlightKind, Disposable, Diagnostic, SignatureHelp} from 'vs/workbench/api/node/extHostTypes';
 import {IPosition, IRange, ISingleEditOperation} from 'vs/editor/common/editorCommon';
 import * as modes from 'vs/editor/common/modes';
 import {CancellationTokenSource} from 'vs/base/common/cancellation';
-import {ExtHostModelService} from 'vs/workbench/api/common/extHostDocuments';
+import {ExtHostModelService} from 'vs/workbench/api/node/extHostDocuments';
 import {IMarkerService, IMarker} from 'vs/platform/markers/common/markers';
-import {ExtHostCommands} from 'vs/workbench/api/common/extHostCommands';
+import {ExtHostCommands} from 'vs/workbench/api/node/extHostCommands';
 import {DeclarationRegistry} from 'vs/editor/contrib/goToDeclaration/common/goToDeclaration';
 import {ExtraInfoRegistry} from 'vs/editor/contrib/hover/common/hover';
 import {OccurrencesRegistry} from 'vs/editor/contrib/wordHighlighter/common/wordHighlighter';
