+++ src/vs/workbench/api/node/extHostStatusBar.ts
