+++ src/vs/workbench/api/node/extHostTelemetry.ts
