+++ src/vs/workbench/api/node/extHostWorkspace.ts
@@ -15,7 +15,7 @@ import {ITextFileService} from 'vs/workbench/parts/files/common/files';
 import {ICommonCodeEditor} from 'vs/editor/common/editorCommon';
 import {bulkEdit, IResourceEdit} from 'vs/editor/common/services/bulkEdit';
 import {TPromise} from 'vs/base/common/winjs.base';
-import {fromRange} from 'vs/workbench/api/common/extHostTypeConverters';
+import {fromRange} from 'vs/workbench/api/node/extHostTypeConverters';
 import {Uri, CancellationToken} from 'vscode';
 
 export class ExtHostWorkspace {
