+++ src/vs/workbench/api/node/mainThreadEditors.ts
